boostrap and firmware updates features wip
This commit moves src/main.py to user/app.py in to conform intended use
case of allowing for firmware downloads from the network.

ssa/bootstrap.py is now the default entry point.
When user code is not present, #12 is executed so this commit closes #12
Note: even though #12 is implemented, #15 functionality is not yet
finished so the feature is not fully ready to merge into main